requirements.txt

streamlit
gtts
Pillow
